requirements.txt
requests==2.27.1
medsenger_api==0.1.20
channels==3.0.4
packaging==21.3
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.12.4
django_cleanup==6.0.0
slack_sdk==3.17.0
sentry-sdk==1.5.12
